requirements.txt
django>=1.9.0
django-crispy-forms
django-reversion
django-formtools==1.0
httplib2==0.9.2
django-xadmin>=0.5.0